requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.18
uvicorn[standard]>=0.34.0
chromadb>=1.0.0
pyyaml>=6.0
